Minor fixes for discord-py-interactions upgrade

+++ src/timer_bot.py
@@ -219,17 +219,18 @@ class Timer:
                     embeds=embeds, components=components
                 )
             # messages older than 1h cannot be edited too much, at some point it fails
-            except interactions.LibraryException as e:
+            except interactions.errors.LibraryException as e:
                 logger.info("Failed to edit message: %s", e)
                 old_message = self.message
                 self.message = await self.channel.send(
                     embeds=embeds, components=components
                 )
-                try:
-                    await old_message.delete()
-                except interactions.LibraryException as e:
-                    logger.info("Failed to delete old message: %s", e)
-                    pass
+                if old_message:
+                    try:
+                        await old_message.delete()
+                    except interactions.errors.LibraryException as e:
+                        logger.info("Failed to delete old message: %s", e)
+                        pass
         else:
             self.message = await self.channel.send(embeds=embeds, components=components)
         if self.thresholds and self.thresholds[-1] >= self.time_left >= 0:
